added deleteDialog for Project in UI  - add delete Button for ProjectTab  - add DeleteDialog in ProjectTab  - extend stateMachine (State: Delete, Event: Delete)  - add new transients (delete, cancel, accept)
 added delete Command in ProjectDto
 extend the ProjectService for deletion (not functional - WIP)

+++ scrum/src/main/groovy/de/fh_zwickau/pti/geobe/dto/UserStoryDto.groovy
@@ -22,7 +22,7 @@ class UserstoryDto {
         Long id = 0
         String name = ''
         String description = ''
-        int priority = 1
+        Long priority = 1
         ProjectDto.QNode project = new ProjectDto.QNode()
         TaskDto.QList backlog = new TaskDto.QList()
 //        def available = new TaskDto.QList()
